+++ build_runner/lib/src/package_graph/package_graph.dart
@@ -12,9 +12,7 @@ import 'dependency_ordering.dart';
 
 /// The SDK package, we filter this to the core libs and dev compiler
 /// resources.
-final PackageNode _sdkPackageNode = new PackageNode(
-    r'$sdk', null, null, getSdkPath(),
-    includes: ['lib/dev_compiler/**.js']);
+final PackageNode _sdkPackageNode = new PackageNode(r'$sdk', getSdkPath());
 
 /// A graph of the package dependencies for an application.
 class PackageGraph {
@@ -90,19 +88,14 @@ class PackageGraph {
 
     /// Create all [PackageNode]s for all deps.
     var nodes = <String, PackageNode>{};
-    Map<String, dynamic> rootDeps;
-    PackageNode addNodeAndDeps(YamlMap yaml, PackageDependencyType type,
-        {bool isRoot: false}) {
+    PackageNode addNodeAndDeps(YamlMap yaml, {bool isRoot: false}) {
       var name = yaml['name'] as String;
       assert(!nodes.containsKey(name));
-      var node = new PackageNode(
-          name, yaml['version'] as String, type, packageLocations[name],
-          includes: isRoot ? ['**'] : ['lib/**']);
+      var node = new PackageNode(name, packageLocations[name]);
       nodes[name] = node;
 
       var deps = _depsFromYaml(yaml, isRoot: isRoot);
-      if (isRoot) rootDeps = deps;
-      deps.forEach((name, source) {
+      for (final name in deps) {
         var dep = nodes[name];
         if (dep == null) {
           var uri = packageLocations[name];
@@ -110,16 +103,15 @@ class PackageGraph {
             throw 'No package found for $name.';
           }
           var pubspec = _pubspecForPath(uri);
-          dep = addNodeAndDeps(pubspec, _dependencyType(rootDeps[name]));
+          dep = addNodeAndDeps(pubspec);
         }
         node.dependencies.add(dep);
-      });
+      }
 
       return node;
     }
 
-    var root =
-        addNodeAndDeps(rootYaml, PackageDependencyType.path, isRoot: true);
+    var root = addNodeAndDeps(rootYaml, isRoot: true);
     return new PackageGraph._(root, nodes);
   }
 
@@ -171,56 +163,17 @@ class PackageNode {
   /// The name of the package as listed in `pubspec.yaml`.
   final String name;
 
-  /// The version of the package as listed in `pubspec.yaml`.
-  ///
-  /// May be `null` if [PackageNode.noPubspec] was used.
-  final String version;
-
-  /// The type of dependency being used to pull in this package.
-  ///
-  /// May be `null` if [PackageNode.noPubspec] was used.
-  final PackageDependencyType dependencyType;
-
   /// All the packages that this package directly depends on.
   final List<PackageNode> dependencies = [];
 
   /// The absolute path of the current version of this package.
   final String path;
 
-  /// The glob patterns to include from this package.
-  ///
-  /// Only the files matching these patterns will end up in the asset graph,
-  /// unless they also match a pattern in [excludes].
-  final List<String> _includes;
-  Iterable<String> get includes => _includes;
-
-  /// The glob patterns to exclude from this package.
-  ///
-  /// Any files matching these pattern will be excluded from the asset graph.
-  final List<String> _excludes;
-  Iterable<String> get excludes => _excludes;
-
-  PackageNode(this.name, this.version, this.dependencyType, String path,
-      {Iterable<String> includes, Iterable<String> excludes})
-      : path = _toAbsolute(path),
-        this._includes = includes?.toList() ?? ['lib/**'],
-        this._excludes = excludes?.toList() ?? <String>[];
-
-  /// Create a [PackageNode] without any details from a `pubspec.yaml` file.
-  ///
-  /// This is useful for testing, or in cases where a package may be synthetic.
-  factory PackageNode.noPubspec(String name,
-          {String path,
-          Iterable<String> includes,
-          Iterable<String> excludes}) =>
-      new PackageNode(name, null, null, _toAbsolute(path),
-          includes: includes, excludes: excludes);
+  PackageNode(this.name, String path) : path = _toAbsolute(path);
 
   @override
   String toString() => '''
   $name:
-    version: $version
-    type: $dependencyType
     path: $path
     dependencies: [${dependencies.map((d) => d.name).join(', ')}]''';
 
@@ -231,38 +184,13 @@ class PackageNode {
   }
 }
 
-/// The type of dependency being used. This dictates how the package should be
-/// watched for changes.
-enum PackageDependencyType { pub, github, path, hosted }
-
-PackageDependencyType _dependencyType(source) {
-  if (source is String || source == null) return PackageDependencyType.pub;
-
-  assert(source is YamlMap);
-  var map = source as YamlMap;
-
-  for (var key in map.keys) {
-    switch (key as String) {
-      case 'git':
-        return PackageDependencyType.github;
-      case 'hosted':
-        return PackageDependencyType.hosted;
-      case 'path':
-      case 'sdk': // Until Flutter supports another type, assume same as path.
-        return PackageDependencyType.path;
-    }
-  }
-  throw 'Unable to determine dependency type:\n$source';
-}
-
 /// Gets the deps from a yaml file, taking into account dependency_overrides.
-Map<String, dynamic> _depsFromYaml(YamlMap yaml, {bool isRoot: false}) {
-  var deps = new Map<String, dynamic>.from(yaml['dependencies'] as Map ?? {});
+Iterable<String> _depsFromYaml(YamlMap yaml, {bool isRoot: false}) {
+  var deps = new Set<String>.from((yaml['dependencies'] as Map)?.keys ?? []);
   if (isRoot) {
-    deps.addAll(new Map.from(yaml['dev_dependencies'] as Map ?? {}));
-    yaml['dependency_overrides']?.forEach((dep, source) {
-      deps[dep as String] = source;
-    });
+    deps.addAll((yaml['dev_dependencies'] as Map<String, dynamic>)?.keys ?? []);
+    deps.addAll(
+        (yaml['dependency_overrides'] as Map<String, dynamic>)?.keys ?? []);
   }
   return deps;
 }